requirements.txt
asgiref==3.3.2
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-templatetag-sugar==1.0
djangorestframework==3.12.4
django-cors-headers>=3.7,<4.0
flake8==3.9.0
gunicorn==20.1.0
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
sqlparse==0.4.1
